Rename layer identity
